Moved common_functions.h to include folder and updated paths
