+++ ImageInLib/ImageInLib/include/common_functions.h
