+++ ImageInLib/ImageInLib/include/registration.h
@@ -3,7 +3,7 @@ extern "C" {
 #endif
 
 #pragma once
-#include "../src/common_functions.h"
+#include "common_functions.h"
 #include "../src/registration_params.h"
 
 	void imageRegistration(dataType **, dataType **, dataType **, const size_t imageHeight, const size_t imageLength, const size_t imageWidth, Registration_Params params, Optimization_Method optimization);
